+++ src/main/java/wooteco/subway/exception/SubwayAdvice.java
@@ -5,6 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import wooteco.subway.exception.auth.InvalidTokenException;
+import wooteco.subway.exception.dto.SubwayExceptionResponse;
 
 @RestControllerAdvice
 public class SubwayAdvice {
