+++ src/main/java/wooteco/subway/exception/dto/SubwayExceptionResponse.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.dto;
 
 public class SubwayExceptionResponse {
     private int httpStatus;
